# Basic_Versioning/Controllers/Company_swagger.yaml
openapi: 3.0.3
info:
  title: Basic_NoVersioning
  description: Title
  version: 1.0.0
servers:
  - url: 'http://localhost:5129'
paths:
  /Company/SetCompanyName:
    put:
      tags:
        - Company
      operationId: SetCompanyName
      description: Set company name and swagger returns it
      parameters:
        - in: query
          name: companyName
          description: set company name
          required: true
          schema:
            type: string
      responses:
        404:
          description: Not found
          content:
            application/json:
              schema:
                type: string
        200:
          description: Success
          content:
            application/json:
              schema:
                type: string
            text/plain:
              schema:
                type: string
...